Move data containers into DTO
