+++ src/backend/app/Jobs/ApplianceTransactionProcessor.php
@@ -2,11 +2,11 @@
 
 namespace App\Jobs;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\TransactionFailedEvent;
 use App\Exceptions\ApplianceTokenNotProcessedException;
 use App\Exceptions\TransactionAmountNotEnoughException;
 use App\Exceptions\TransactionNotInitializedException;
-use App\Misc\TransactionDataContainer;
 use App\Models\Transaction\Transaction;
 use App\Utils\ApplianceInstallmentPayer;
 use Illuminate\Support\Facades\Log;
