+++ src/backend/app/Jobs/TokenProcessor.php
@@ -2,10 +2,10 @@
 
 namespace App\Jobs;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\PaymentSuccessEvent;
 use App\Events\TransactionFailedEvent;
 use App\Events\TransactionSuccessfulEvent;
-use App\Misc\TransactionDataContainer;
 use App\Models\AssetRate;
 use App\Models\Token;
 use Illuminate\Bus\Queueable;
