+++ src/backend/app/Services/AppliancePaymentService.php
@@ -2,11 +2,11 @@
 
 namespace App\Services;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\NewLogEvent;
 use App\Events\PaymentSuccessEvent;
 use App\Exceptions\PaymentAmountBiggerThanTotalRemainingAmount;
 use App\Exceptions\PaymentAmountSmallerThanZero;
-use App\Misc\TransactionDataContainer;
 use App\Models\AssetPerson;
 use App\Models\AssetRate;
 use App\Models\Device;
