+++ src/backend/app/Utils/ApplianceInstallmentPayer.php
@@ -2,9 +2,9 @@
 
 namespace App\Utils;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\PaymentSuccessEvent;
 use App\Exceptions\Device\DeviceIsNotAssignedToCustomer;
-use App\Misc\TransactionDataContainer;
 use App\Models\AssetPerson;
 use App\Models\Device;
 use App\Models\Person\Person;
