+++ src/backend/packages/inensus/angaza-shs/src/Modules/Api/AngazaSHSApi.php
@@ -2,10 +2,10 @@
 
 namespace Inensus\AngazaSHS\Modules\Api;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\NewLogEvent;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use Carbon\Carbon;
