+++ src/backend/packages/inensus/chint-meter/src/Modules/Api/ChintMeterApi.php
@@ -2,9 +2,9 @@
 
 namespace Inensus\ChintMeter\Modules\Api;
 
+use App\DTO\TransactionDataContainer;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use Illuminate\Support\Facades\Log;
 use Inensus\ChintMeter\Exceptions\TariffPriceDoesNotMatchException;
