+++ src/backend/packages/inensus/daly-bms/src/Modules/Api/DalyBmsApi.php
@@ -2,10 +2,10 @@
 
 namespace Inensus\DalyBms\Modules\Api;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\NewLogEvent;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use App\Models\User;
