+++ src/backend/packages/inensus/demo-meter-manufacturer/src/DemoMeterManufacturerApi.php
@@ -2,8 +2,8 @@
 
 namespace Inensus\DemoMeterManufacturer;
 
+use App\DTO\TransactionDataContainer;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use Inensus\DemoMeterManufacturer\Models\DemoMeterTransaction;
