+++ src/backend/packages/inensus/demo-shs-manufacturer/src/DemoShsManufacturerApi.php
@@ -2,10 +2,10 @@
 
 namespace Inensus\DemoShsManufacturer;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\NewLogEvent;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use Inensus\DemoShsManufacturer\Models\DemoShsTransaction;
