+++ src/backend/packages/inensus/spark-meter/src/SparkMeterApi.php
@@ -2,9 +2,9 @@
 
 namespace Inensus\SparkMeter;
 
+use App\DTO\TransactionDataContainer;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use GuzzleHttp\Client;
