+++ src/backend/packages/inensus/stron-meter/src/StronMeterApi.php
@@ -2,9 +2,9 @@
 
 namespace Inensus\StronMeter;
 
+use App\DTO\TransactionDataContainer;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\MainSettings;
 use App\Models\Token;
