+++ src/backend/packages/inensus/sun-king-shs/src/Modules/Api/SunKingSHSApi.php
@@ -2,10 +2,10 @@
 
 namespace Inensus\SunKingSHS\Modules\Api;
 
+use App\DTO\TransactionDataContainer;
 use App\Events\NewLogEvent;
 use App\Exceptions\Manufacturer\ApiCallDoesNotSupportedException;
 use App\Lib\IManufacturerAPI;
-use App\Misc\TransactionDataContainer;
 use App\Models\Device;
 use App\Models\Token;
 use Illuminate\Support\Facades\Log;
