+++ src/backend/routes/web.php
@@ -11,10 +11,10 @@
  * |.
  */
 
+use App\DTO\TransactionDataContainer;
 use App\Events\TransactionSuccessfulEvent;
 use App\Jobs\EnergyTransactionProcessor;
 use App\Jobs\TokenProcessor;
-use App\Misc\TransactionDataContainer;
 use App\Models\Transaction\Transaction;
 use Illuminate\Support\Facades\Route;
